Added filtering of empty extra properties
Keystone doesn't allow to delete extra properties. We can unset
their values: the same operation as to set property value to None
(according to action presented in V2), but keystone will continue
store these properties in DB and return them with None values.
This behavior significantly influences on UX, when user activelly
uses these fields to mark/unmark project (users, roles, domains),
because it leads to mess in an output of CLI or UI.

This patch adds filtering of empty properties to prevent cluttering
of extra field that contains custom properties for objects.

Change-Id: Idb79e422f6cf9743e37c78abdfd6a13869ef17c0